fix(agent): tighten message dispatch diagnostics
Keep dispatch lifecycle code within repository size limits and report MCP attachment failures with an accurate terminal error code.
